ecr-credential-helper doesn't support chained role assuming
This is the reason I was doing the assume role in the agent python code and
handing temp access key/secret over to bootstrap: in that situation, the root
creds are real creds. In this new approach, the root creds are ECS task role, so
it's a long chain of roles with no "real" credentials: ECS Role -> MM Role -> MM
Access Role -- somewhere in the dependencies of aws-ecr-credential-helper (I
think go-aws-sdk, though it looks like they released a fix for it), there's a
limitation in this respect. I'm hoping since aws ecr get-login just does the
assume-role immediately and gets the hard credentials, this should work around
the issue while still fixing the MM problem.